@@ -3,20 +3,20 @@
   "version": "5.0.1",
   "description": "",
   "dependencies": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "cordova-plugin-splashscreen": "6.0.0",
     "monaca-plugin-monaca-core": "3.3.1",
-    "onsenui": "2.11.2",
-    "react": "^16.8.6",
-    "react-dom": "^16.4.1",
-    "react-onsenui": "1.11.4"
+    "onsenui": "~2.12.2",
+    "react-onsenui": "~1.13.0"
   },
   "scripts": {
-    "monaca:preview": "npm run dev & npm run watch",
+    "monaca:preview": "npm run dev",
     "monaca:transpile": "npm run build",
     "monaca:debug": "npm run watch",
-    "dev": "webpack serve --open",
-    "build": "webpack --mode production",
-    "watch": "webpack --watch --mode production"
+    "dev": "cross-env NODE_ENV=development vite",
+    "watch": "cross-env NODE_ENV=development vite build --watch",
+    "build": "cross-env NODE_ENV=production vite build"
   },
   "cordova": {
     "plugins": {
@@ -25,30 +25,12 @@
     }
   },
   "devDependencies": {
-    "@babel/core": "^7.17.5",
-    "@babel/preset-env": "^7.16.11",
-    "@babel/preset-react": "^7.16.7",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^8.2.3",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "cordova": "^11.0.0",
-    "css-loader": "^6.6.0",
-    "cssnano": "^5.1.0",
-    "file-loader": "^6.2.0",
-    "html-loader": "^3.1.0",
-    "html-webpack-plugin": "^5.5.0",
-    "mini-css-extract-plugin": "^2.5.3",
-    "postcss": "^8.4.7",
-    "postcss-loader": "^6.2.1",
-    "postcss-preset-env": "^7.4.1",
-    "postcss-url": "^10.1.3",
-    "progress-bar-webpack-plugin": "^2.1.0",
-    "react-hot-loader": "^4.13.0",
-    "style-loader": "^3.3.1",
-    "webpack": "^5.69.1",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.7.4",
-    "yargs": "^17.3.1"
+    "cross-env": "^7.0.3",
+    "@types/react": "^18.0.17",
+    "@types/react-dom": "^18.0.6",
+    "@vitejs/plugin-react": "^2.0.1",
+    "postcss-preset-env": "^7.4.2",
+    "vite": "^3.0.7",
+    "yargs": "^17.4.1"
   }
 }
