@@ -1,10 +1,5 @@
 module.exports = {
-  plugins: [
-    [
-      "postcss-preset-env",
-      {
-        // Options
-      },
-    ],
-  ],
+  plugins: {
+    'postcss-preset-env': {},
+  },
 };
