@@ -1,11 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import {Page, Button, Toolbar} from 'react-onsenui';
-import {notification} from 'onsenui';
+import ons from 'onsenui';
 
 export default class App extends React.Component {
   alertPopup() {
-    notification.alert('This is an Onsen UI alert notification test.');
+    ons.notification.alert('This is an Onsen UI alert notification test.');
   }
 
   renderToolbar() {
